fix: #657 use only one vega tx dialog
