--- libs/wallet/src/vega-order-transaction-dialog/index.ts
@@ -1,1 +1,0 @@
-export * from './vega-order-transaction-dialog';
